backend/requirements.txt
# ============================================================
# Adobe Hackathon Backend - Optimized for Render Free Tier
# ============================================================
# Removed: torch, sentence-transformers, scikit-learn (heavy packages)
# Using: API-based embeddings from OpenAI/Google instead
# ============================================================

# Core FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
python-multipart==0.0.6
python-dotenv==1.0.0

# PDF processing dependencies
PyPDF2==3.0.1
pymupdf==1.23.8

# Lightweight vector database
faiss-cpu==1.7.4
numpy==1.24.3

# LLM API dependencies (essential only)
openai==1.3.7
google-generativeai==0.3.2
anthropic==0.7.8

# LangChain for RAG (use older stable versions)
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.5
langchain-community==0.0.10

# HTTP and utility dependencies
requests==2.31.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Azure services for TTS
azure-cognitiveservices-speech==1.31.0

# Audio processing
pydub==0.25.1

# Database
sqlalchemy==2.0.23

# Security
python-jose[cryptography]==3.3.0

# Supabase
supabase==2.3.4

# Sentence Transformers for RAG
sentence-transformers==2.2.2

# Google Generative AI (direct library, not langchain wrapper)
google-generativeai>=0.3.0

# CORS support
python-multipart==0.0.6
